implement review findSingleItem route

+++ src/api/v1/book/controllers/removeItem.js
@@ -1,7 +1,11 @@
 const bookService=require('../../../../lib/book')
 const removeItem=async(req,res,next)=>{
     const {id}=req.params
-
+    if(!id) throw error('Id is required',400)
+    
+    const isValid=isValidObjectId(id)
+    if(!isValid) throw error('Invalid id format',400)
+        
     try{
         await bookService.removeItem({id})
         res.status(204).end()
